Fix/disable tests
RISDEV-4317
